train-test split methodology in readme

+++ JGNN/src/main/java/mklab/JGNN/data/IdConverter.java
@@ -15,51 +15,51 @@ import mklab.JGNN.nn.operations.Concat;
  * @author Emmanouil Krasanakis
  */
 public class IdConverter {
-	protected HashMap<Object, Integer> ids = new HashMap<Object, Integer>();
-	protected HashMap<Integer, Object> inverse = new HashMap<Integer, Object>();
+	protected HashMap<Object, Long> ids = new HashMap<Object, Long>();
+	protected HashMap<Long, Object> inverse = new HashMap<Long, Object>();
 	public IdConverter() {
 	}
 	public IdConverter(ArrayList<?> objects) {
 		for(int i=0;i<objects.size();i++)
 			getOrCreateId(objects.get(i));
 	}
-	public int getOrCreateId(Object object) {
-		Integer ret = ids.get(object);
+	public long getOrCreateId(Object object) {
+		Long ret = ids.get(object);
 		if(ret==null) {
-			ids.put(object, ret = ids.size());
+			ids.put(object, ret = (long)ids.size());
 			inverse.put(ret, object);
 		}
 		return ret;
 	}
-	public Object get(int id) {
+	public Object get(long id) {
 		return inverse.get(id);
 	}
-	public int getId(Object object) {
+	public long getId(Object object) {
 		return ids.get(object);
 	}
-	public int size() {
+	public long size() {
 		return ids.size();
 	}
 	public boolean contains(Object object) {
 		return ids.containsKey(object);
 	}
-	public ArrayList<Integer> getIds() {
-		return new ArrayList<Integer>(ids.values());
+	public ArrayList<Long> getIds() {
+		return new ArrayList<Long>(ids.values());
 	}
-	public Matrix oneHot(ArrayList<HashMap<Integer, String>> nodeFeatures) {
+	public Matrix oneHot(ArrayList<HashMap<Long, String>> nodeFeatures) {
 		Matrix features = null;
-		for(HashMap<Integer, String> feature : nodeFeatures) {
+		for(HashMap<Long, String> feature : nodeFeatures) {
 			Matrix onehot = oneHot(feature);
 			features = features==null?onehot:(Matrix)new Concat().run(features, onehot);
 		}
 		return features;
 	}
-	public Matrix oneHot(HashMap<Integer, String> nodeLabels) {
+	public Matrix oneHot(HashMap<Long, String> nodeLabels) {
 		IdConverter encoder = new IdConverter();
 		for(String label : nodeLabels.values())
 			encoder.getOrCreateId(label);
 		Matrix encoding = new SparseMatrix(nodeLabels.size(), encoder.size());
-		for(int i=0;i<size();i++)
+		for(long i=0;i<size();i++)
 			encoding.put(i, encoder.getId(nodeLabels.get(i)), 1);
 		return encoding;
 	}
@@ -87,15 +87,15 @@ public class IdConverter {
 			ret.put(i, encoder.getId(nodeLabels.get(i)));
 		return ret;
 	}
-	public List<Tensor> toMultipleTensors(HashMap<Integer, String> nodeLabels) {
+	public List<Tensor> toMultipleTensors(HashMap<Long, String> nodeLabels) {
 		List<Tensor> ret = new ArrayList<Tensor>();
 		IdConverter encoder = new IdConverter();
 		for(String label : nodeLabels.values()) 
 			encoder.getOrCreateId(label);
 		for(int i=0;i<encoder.size();i++)
 			ret.add(new DenseTensor(size()));
-		for(int i=0;i<size();i++)
-			ret.get(encoder.getId(nodeLabels.get(i))).put(i, 1);
+		for(long i=0;i<size();i++)
+			ret.get((int)encoder.getId(nodeLabels.get(i))).put(i, 1);
 		return ret;
 	}
 }
